feat 注释修改

+++ dptms-file-manage/src/main/java/com/darren1112/dptms/file/service/impl/FileCenterServiceImpl.java
@@ -1,13 +1,17 @@
 package com.darren1112.dptms.file.service.impl;
 
+import com.darren1112.dptms.common.core.exception.BadRequestException;
 import com.darren1112.dptms.common.core.util.FileUtil;
 import com.darren1112.dptms.common.spi.file.dto.FileCenterDto;
+import com.darren1112.dptms.file.common.enums.FileManageErrorCodeEnum;
 import com.darren1112.dptms.file.repository.FileCenterRepository;
 import com.darren1112.dptms.file.service.FileCenterService;
+import com.darren1112.dptms.file.service.FileInfoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CacheConfig;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.Cacheable;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -17,7 +21,7 @@ import java.util.Optional;
 /**
  * 文件中心ServiceImpl
  *
- * @author luyuhao
+ * @author darren
  * @since 2021/12/18
  */
 @Service
@@ -28,12 +32,16 @@ public class FileCenterServiceImpl implements FileCenterService {
     @Autowired
     private FileCenterRepository fileCenterRepository;
 
+    @Lazy
+    @Autowired
+    private FileInfoService fileInfoService;
+
     /**
      * 根据父节点id查询
      *
      * @param parentId 父节点id
      * @return {@link FileCenterDto}
-     * @author luyuhao
+     * @author darren
      * @since 2021/12/18
      */
     @Override
@@ -57,7 +65,7 @@ public class FileCenterServiceImpl implements FileCenterService {
      * 新增文件/文件夹
      *
      * @param dto 文件/文件夹信息
-     * @author luyuhao
+     * @author darren
      * @since 2021/12/19
      */
     @Override
@@ -78,6 +86,14 @@ public class FileCenterServiceImpl implements FileCenterService {
     @CacheEvict(allEntries = true)
     @Transactional(rollbackFor = Throwable.class)
     public void update(FileCenterDto dto) {
+        FileCenterDto dataInfo = fileCenterRepository.getById(dto.getId());
+        if (dataInfo == null) {
+            throw new BadRequestException(FileManageErrorCodeEnum.FILE_NOT_EXIST);
+        }
+        if (Integer.valueOf(1).equals(dataInfo.getFileType())) {
+            // 同步更新文件信息
+            fileInfoService.updateFileName(dto.getFileId(), dto.getFileName(), dto.getUpdater());
+        }
         fileCenterRepository.updateInfo(dto);
     }
 
@@ -92,7 +108,12 @@ public class FileCenterServiceImpl implements FileCenterService {
     @CacheEvict(allEntries = true)
     @Transactional(rollbackFor = Throwable.class)
     public void deleteById(FileCenterDto dto) {
+        FileCenterDto dataInfo = fileCenterRepository.getById(dto.getId());
+        if (dataInfo == null) {
+            throw new BadRequestException(FileManageErrorCodeEnum.FILE_NOT_EXIST);
+        }
         fileCenterRepository.deleteById(dto);
         // TODO 删除文件/文件夹下的所有文件
+        List<FileCenterDto> subFIleList = fileCenterRepository.listSubFileList(dataInfo.getFileParentPath()+ "/" + dataInfo.getId());
     }
 }
